+++ packages/saas-ui-date-picker/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @saas-ui/date-picker
 
+## 0.12.42
+
+### Patch Changes
+
+- @saas-ui/modals@2.1.40
+
+## 0.12.41
+
+### Patch Changes
+
+- 5db3102: fix(date-picker): expose useDatePickerInput and useDateRangePicker hooks
+
 ## 0.12.40
 
 ### Patch Changes