+++ packages/storybook/CHANGELOG.md
@@ -1,5 +1,12 @@
 # saas-ui-storybook
 
+## 2.0.10
+
+### Patch Changes
+
+- @saas-ui/react@2.8.10
+- @saas-ui/storybook-addon@3.0.2
+
 ## 2.0.9
 
 ### Patch Changes